// admin-client/src/rest/gateway/models/EmployeeDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gateway Service API
 * API for Gateway
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface EmployeeDto
 */
export interface EmployeeDto {
    /**
     * 
     * @type {number}
     * @memberof EmployeeDto
     */
    age?: number;
    /**
     * 
     * @type {number}
     * @memberof EmployeeDto
     */
    departmentId?: number;
    /**
     * 
     * @type {number}
     * @memberof EmployeeDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof EmployeeDto
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof EmployeeDto
     */
    organizationId?: number;
    /**
     * 
     * @type {string}
     * @memberof EmployeeDto
     */
    position?: string;
}

export function EmployeeDtoFromJSON(json: any): EmployeeDto {
    return EmployeeDtoFromJSONTyped(json, false);
}

export function EmployeeDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): EmployeeDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'age': !exists(json, 'age') ? undefined : json['age'],
        'departmentId': !exists(json, 'departmentId') ? undefined : json['departmentId'],
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'organizationId': !exists(json, 'organizationId') ? undefined : json['organizationId'],
        'position': !exists(json, 'position') ? undefined : json['position'],
    };
}

export function EmployeeDtoToJSON(value?: EmployeeDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'age': value.age,
        'departmentId': value.departmentId,
        'id': value.id,
        'name': value.name,
        'organizationId': value.organizationId,
        'position': value.position,
    };
}